+++ src/setup-docker/templates/fabric-docker/commands-generated.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 
-function generateArtifacts() {
+generateArtifacts() {
   printHeadline "Generating basic configs" "U1F913"
 
-  <% orgs.forEach(function(org){ -%>
+  <% orgs.forEach((org) => { -%>
     printItalics "Generating crypto material for <%= org.name %>" "U1F512"
     certsGenerate <% -%>
       "$FABLO_NETWORK_ROOT/fabric-config" <% -%>
@@ -12,7 +12,7 @@ function generateArtifacts() {
       "$FABLO_NETWORK_ROOT/fabric-config/crypto-config/"
 
   <% }) -%>
-  <%_ ordererGroups.forEach(function(ordererGroup) { _%>
+  <%_ ordererGroups.forEach((ordererGroup) => { _%>
   printItalics "Generating genesis block for group <%= ordererGroup.name %>" "U1F3E0"
   genesisBlockCreate "$FABLO_NETWORK_ROOT/fabric-config" "$FABLO_NETWORK_ROOT/fabric-config/config" "<%= ordererGroup.profileName %>"
 
@@ -21,30 +21,30 @@ function generateArtifacts() {
   mkdir -p "$FABLO_NETWORK_ROOT/fabric-config/chaincode-packages"
 }
 
-function startNetwork() {
+startNetwork() {
   printHeadline "Starting network" "U1F680"
   (cd "$FABLO_NETWORK_ROOT"/fabric-docker && docker-compose up -d)
   sleep 4
 }
 
-function generateChannelsArtifacts() {
+generateChannelsArtifacts() {
   <% if (!channels || !channels.length) { -%>
     echo "No channels"
   <% } else { -%>
-    <% channels.forEach(function(channel){  -%>
+    <% channels.forEach((channel) => {  -%>
       printHeadline "Generating config for '<%= channel.name %>'" "U1F913"
       createChannelTx "<%= channel.name %>" "$FABLO_NETWORK_ROOT/fabric-config" "<%= channel.profileName %>" "$FABLO_NETWORK_ROOT/fabric-config/config"
     <% }) -%>
   <% } -%>
 }
 
-function installChannels() {
+installChannels() {
   <% if (!channels || !channels.length) { -%>
     echo "No channels"
   <% } else { -%>
-    <% channels.forEach(function(channel){ -%>
-      <% channel.orgs.forEach(function(org, orgNo){ -%>
-        <% org.peers.forEach(function(peer, peerNo){ -%>
+    <% channels.forEach((channel) => { -%>
+      <% channel.orgs.forEach((org, orgNo) => { -%>
+        <% org.peers.forEach((peer, peerNo) => { -%>
           <% if(orgNo == 0 && peerNo == 0) { -%>
             printHeadline "Creating '<%= channel.name %>' on <%= org.name %>/<%= peer.name %>" "U1F63B"
             <% if(!networkSettings.tls) { -%>
@@ -70,11 +70,11 @@ function installChannels() {
   <% } -%>
 }
 
-function installChaincodes() {
+installChaincodes() {
   <% if (!chaincodes || !chaincodes.length) { -%>
     echo "No chaincodes"
   <% } else { -%>
-    <% chaincodes.forEach(function(chaincode) { -%>
+    <% chaincodes.forEach((chaincode) => { -%>
       if [ -n "$(ls "$CHAINCODES_BASE_DIR/<%= chaincode.directory %>")" ]; then
         <% if (networkSettings.capabilities.isV2) { -%>
           local version="<%= chaincode.version %>"
@@ -90,10 +90,10 @@ function installChaincodes() {
   <% } -%>
 }
 
-function notifyOrgsAboutChannels() {
+notifyOrgsAboutChannels() {
   printHeadline "Creating new channel config blocks" "U1F537"
-  <% channels.forEach(function(channel){ -%>
-    <% channel.orgs.forEach(function(org){ -%>
+  <% channels.forEach((channel) => { -%>
+    <% channel.orgs.forEach((org) => { -%>
       createNewChannelUpdateTx <% -%>
         "<%= channel.name %>" <% -%>
         "<%= org.mspName %>" <% -%>
@@ -104,8 +104,8 @@ function notifyOrgsAboutChannels() {
   <% }) %>
 
   printHeadline "Notyfing orgs about channels" "U1F4E2"
-  <% channels.forEach(function(channel){ -%>
-    <% channel.orgs.forEach(function(org){ -%>
+  <% channels.forEach((channel) => { -%>
+    <% channel.orgs.forEach((org) => { -%>
      <% if(!networkSettings.tls) { -%>
        notifyOrgAboutNewChannel <% -%>
          "<%= channel.name %>" <% -%>
@@ -126,21 +126,21 @@ function notifyOrgsAboutChannels() {
   <% }) %>
 
   printHeadline "Deleting new channel config blocks" "U1F52A"
-  <% channels.forEach(function(channel){ -%>
-    <% channel.orgs.forEach(function(org){ -%>
+  <% channels.forEach((channel) => { -%>
+    <% channel.orgs.forEach((org) => { -%>
       deleteNewChannelUpdateTx "<%= channel.name %>" "<%= org.mspName %>" "<%= org.cli.address %>"
     <% }) -%>
   <% }) -%>
 }
 
-function upgradeChaincode() {
+upgradeChaincode() {
   local chaincodeName="$1"
   if [ -z "$chaincodeName" ]; then echo "Error: chaincode name is not provided"; exit 1; fi
 
   local version="$2"
   if [ -z "$version" ]; then echo "Error: chaincode version is not provided"; exit 1; fi
 
-  <% chaincodes.forEach(function(chaincode) { -%>
+  <% chaincodes.forEach((chaincode) => { -%>
     if [ "$chaincodeName" = "<%= chaincode.name %>" ]; then
       if [ -n "$(ls "$CHAINCODES_BASE_DIR/<%= chaincode.directory %>")" ]; then
         <% if (networkSettings.capabilities.isV2) { -%>
@@ -156,20 +156,20 @@ function upgradeChaincode() {
   <% }) -%>
 }
 
-function stopNetwork() {
+stopNetwork() {
   printHeadline "Stopping network" "U1F68F"
   (cd "$FABLO_NETWORK_ROOT"/fabric-docker && docker-compose stop)
   sleep 4
 }
 
-function networkDown() {
+networkDown() {
   printHeadline "Destroying network" "U1F916"
   (cd "$FABLO_NETWORK_ROOT"/fabric-docker && docker-compose down)
 
   printf "\nRemoving chaincode containers & images... \U1F5D1 \n"
-  <% chaincodes.forEach(function(chaincode) { -%>
-    <% chaincode.channel.orgs.forEach(function (org) { -%>
-      <% org.peers.forEach(function (peer) { -%>
+  <% chaincodes.forEach((chaincode) => { -%>
+    <% chaincode.channel.orgs.forEach((org) => { -%>
+      <% org.peers.forEach((peer) => { -%>
         <% const chaincodeContainerName=`dev-${peer.address}-${chaincode.name}-${chaincode.version}` -%>
         docker rm -f $(docker ps -a | grep <%= chaincodeContainerName %>-* | awk '{print $1}') || echo "docker rm failed, Check if all fabric dockers properly was deleted"
         docker rmi $(docker images <%= chaincodeContainerName %>-* -q) || echo "docker rm failed, Check if all fabric dockers properly was deleted"