+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Chore & Maintenance
 * Command 'reboot' renamed to 'reset'
+* Keyword `function` removed from scripts for better portability
 
 ## 0.3.0